menuglyphs - make colours render properly.
Two issues resolved -
1. Object needs to be positioned so mapglyph() thinks the hero can
see it. Solution: pass u.ux,u.uy to mapglyph(). This is elegant,
and works well - even taking into account blindness (untested).
2. Object materials need to be accounted for. "Solution:" overload
the attr arg to indicate the 'any' opaque type actually contains an
object, pass said object to mapglyph so material can be extracted.
This is a horrible kludge, which due to the overloading of the attr
flag may break other naiive windowports.  Do not copy this
implementation, it is very bad.